back_end/microserviceGIR/src/main/webapp/app/entities/microserviceGIR/baccalaureat/route/baccalaureat-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IBaccalaureat } from '../baccalaureat.model';
import { BaccalaureatService } from '../service/baccalaureat.service';

import baccalaureatResolve from './baccalaureat-routing-resolve.service';

describe('Baccalaureat routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let service: BaccalaureatService;
  let resultBaccalaureat: IBaccalaureat | null | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    service = TestBed.inject(BaccalaureatService);
    resultBaccalaureat = undefined;
  });

  describe('resolve', () => {
    it('should return IBaccalaureat returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      TestBed.runInInjectionContext(() => {
        baccalaureatResolve(mockActivatedRouteSnapshot).subscribe({
          next(result) {
            resultBaccalaureat = result;
          },
        });
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultBaccalaureat).toEqual({ id: 123 });
    });

    it('should return null if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      TestBed.runInInjectionContext(() => {
        baccalaureatResolve(mockActivatedRouteSnapshot).subscribe({
          next(result) {
            resultBaccalaureat = result;
          },
        });
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultBaccalaureat).toEqual(null);
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse<IBaccalaureat>({ body: null })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      TestBed.runInInjectionContext(() => {
        baccalaureatResolve(mockActivatedRouteSnapshot).subscribe({
          next(result) {
            resultBaccalaureat = result;
          },
        });
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultBaccalaureat).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
